simulation/requirements.txt
Django==4.2.5
psycopg2-binary==2.9.7
djangorestframework==3.14.0
requests==2.28.1
